requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==7.0.0
certifi==2026.1.4
click==8.3.1
fastapi==0.128.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.8.0
pypdf==6.6.2
python-dotenv==1.2.1
python-multipart==0.0.22
sniffio==1.3.1
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
zhipuai==2.1.5.20250825
